(refac + test): Method-shape op trait; Constant scalar/batch consistency
`_output_eltype` previously enumerated standard numeric types via the
`_PromotableValue` union and hand-applied an `Int→Float` upgrade — a manual
proxy for "the arithmetic kernels divide". Selection kernels (Constant) don't
divide, so the upgrade was wrong for them and was masked by sample-first
allocators (one extra scalar eval per batch).

Add a method-shape overload that lets the kernel itself drive inference:

    @inline _output_eltype(kernel_op::F, ::Type{Tv}, types::Type...) where {F, Tv} =
        ... Base.promote_op(kernel_op, Tv, types...) ...

Constant declares its kernel shape and routes its trait through the overload:

    @inline _constant_kernel_shape(yv, q) = yv * one(q)
    @inline _output_eltype(::ConstantInterpolant{Tg, Tv}, ::Type{Tq}) where {Tg, Tv, Tq} =
        _output_eltype(_constant_kernel_shape, Tv, Tq)

Julia infers the exact kernel return type — no `_PromotableValue` enumeration,
no Float upgrade. Int×Int×Int matches the scalar callable (Int), SVector ×
Dual still resolves to SVector{Dual}. Constant ND mirrors the override. The
`_PromotableValue` definition stays (other paths — grid promotion, autocache
gating, FillExtrap auto-convert — use it independently).

Other methods (Linear/Cubic/Quadratic/Hermite/etc.) keep the existing trait —
their kernels divide, so the Float upgrade matches reality. Migrating them to
their own kernel shape op is a follow-up.

Tests:
- Constant Int contract reverted: `itp([2,5,8]) isa Vector{Int}` again,
  matching `itp(3) === 30`. No scalar/batch asymmetry.
- New `@testitem "Type stability — duck-Tv × duck-Tq inference"`: `@inferred`
  coverage for 1D persistent/oneshot × 5 methods (scalar + batch), ND
  persistent/oneshot × 3 methods, ND deriv-zero per-axis carrier, Hermite
  mixed eltype, CubicSeriesInterpolant persistent, anchored-vector callable.
  Catches silent Union returns and `Any` fallbacks `isa` checks would miss.

+++ test/test_constant.jl
@@ -195,10 +195,10 @@ end
         @test out ≈ [10.0, 20.0, 30.0]
 
         # Test 3: Vector allocating with Integer grid + Integer query.
-        # Natural promotion: trait upgrades Int→Float at allocation.
+        # Kernel-shape trait keeps Int×Int×Int in Int; matches scalar path.
         result_vec = constant_interp(x_int, y_int, [0, 1, 2])
-        @test result_vec isa Vector{Float64}
-        @test result_vec == [10.0, 20.0, 30.0]
+        @test result_vec isa Vector{Int}
+        @test result_vec == [10, 20, 30]
 
         # Test 4: In-place Real→Float wrapper (lines 440-468)
         # constant_interp!(output, x::AbstractVector{T}, y::AbstractVector{T}, x_targets::AbstractVector{S})
